made the while loop run - semi base working of calling functions, reading and skipping

+++ src/corewar/includes/vm.h
@@ -77,9 +77,9 @@ typedef struct s_carriage
 	uint8_t				*home;
 	uint8_t				*current;
 	int32_t				skip;//nbr of bytes that needs to be skipped to go to next statement
-	int32_t				registry[REG_NUMBER];//not acccurate numbers//first r1 will be identification number of player on whose code the carraige stands
+	int64_t				registry[REG_NUMBER];//not acccurate numbers//first r1 will be identification number of player on whose code the carraige stands
 	int32_t				arg_types[ARGS];
-	int32_t				args_found[ARGS];
+	int64_t				args_found[ARGS];
 	struct s_carriage	*next;
 }						t_carriage;
